backend/requirements.txt
asgiref==3.8.1
Django==5.0.14
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
mssql-django==1.5
packaging==25.0
pyodbc==5.2.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
django-cors-headers